Separate out Serializable-related code into partial classes for System.Drawing.Common.
