+++ src/System.Drawing.Common/src/System/Drawing/Icon.Serializable.cs
@@ -1,0 +1,51 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Drawing
+{
+    using System.IO;
+    using System.Runtime.Serialization;
+    using System.Security.Permissions;
+
+    [Serializable]
+    public sealed partial class Icon : ISerializable
+    {
+        /// <summary>
+        /// Constructor used in deserialization
+        /// </summary>
+        private Icon(SerializationInfo info, StreamingContext context)
+        {
+            _iconData = (byte[])info.GetValue("IconData", typeof(byte[]));
+            _iconSize = (Size)info.GetValue("IconSize", typeof(Size));
+
+            if (_iconSize.IsEmpty)
+            {
+                Initialize(0, 0);
+            }
+            else
+            {
+                Initialize(_iconSize.Width, _iconSize.Height);
+            }
+        }
+
+        /// <summary>
+        /// ISerializable private implementation
+        /// </summary>
+        [SecurityPermissionAttribute(SecurityAction.LinkDemand, Flags = SecurityPermissionFlag.SerializationFormatter)]
+        void ISerializable.GetObjectData(SerializationInfo si, StreamingContext context)
+        {
+            if (_iconData != null)
+            {
+                si.AddValue("IconData", _iconData, typeof(byte[]));
+            }
+            else
+            {
+                MemoryStream stream = new MemoryStream();
+                Save(stream);
+                si.AddValue("IconData", stream.ToArray(), typeof(byte[]));
+            }
+            si.AddValue("IconSize", _iconSize, typeof(Size));
+        }
+    }
+}
