+++ src/System.Drawing.Common/src/System/Drawing/Printing/InvalidPrinterException.cs
@@ -4,9 +4,7 @@
 
 namespace System.Drawing.Printing
 {
-    using System.Runtime.Serialization;
     using System.Security;
-    using System.Security.Permissions;
 
     /// <include file='doc\InvalidPrinterException.uex' path='docs/doc[@for="InvalidPrinterException"]/*' />
     /// <devdoc>
@@ -16,8 +14,7 @@ namespace System.Drawing.Printing
     ///       exception that is thrown when trying to access a printer using invalid printer settings.
     ///    </para>
     /// </devdoc>
-    [Serializable()]
-    public class InvalidPrinterException : SystemException
+    public partial class InvalidPrinterException : SystemException
     {
         private PrinterSettings _settings;
 
@@ -31,24 +28,6 @@ namespace System.Drawing.Printing
         : base(GenerateMessage(settings))
         {
             _settings = settings;
-        }
-
-        /// <include file='doc\InvalidPrinterException.uex' path='docs/doc[@for="InvalidPrinterException.InvalidPrinterException1"]/*' />
-        protected InvalidPrinterException(SerializationInfo info, StreamingContext context) : base(info, context)
-        {
-            _settings = (PrinterSettings)info.GetValue("settings", typeof(PrinterSettings));
-        }
-
-        /// <include file='doc\InvalidPrinterException.uex' path='docs/doc[@for="InvalidPrinterException.GetObjectData"]/*' />
-        [SecurityPermissionAttribute(SecurityAction.Demand, SerializationFormatter = true)]
-        public override void GetObjectData(SerializationInfo info, StreamingContext context)
-        {
-            if (info == null)
-            {
-                throw new ArgumentNullException("info");
-            }
-            info.AddValue("settings", _settings);
-            base.GetObjectData(info, context);
         }
 
         private static string GenerateMessage(PrinterSettings settings)
