+++ src/System.Drawing.Common/src/System/Drawing/Printing/PaperSize.Serializable.cs
@@ -1,0 +1,11 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Drawing.Printing
+{
+    [Serializable]
+    public partial class PaperSize
+    {
+    }
+}
