+++ src/System.Drawing.Common/src/System/Drawing/Printing/PrinterSettings.cs
@@ -18,8 +18,7 @@ namespace System.Drawing.Printing
     ///    Information about how a document should be printed, including which printer
     ///    to print it on.
     /// </devdoc>
-    [Serializable]
-    public class PrinterSettings : ICloneable
+    public partial class PrinterSettings : ICloneable
     {
         // All read/write data is stored in managed code, and whenever we need to call Win32,
         // we create new DEVMODE and DEVNAMES structures.  We don't store device capabilities,
